backend/requirement.txt
absl-py==2.2.2
asgiref==3.8.1
astunparse==1.6.3
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
flatbuffers==25.2.10
gast==0.4.0
google-auth==2.39.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.71.0
gunicorn==23.0.0
h5py==3.13.0
idna==3.10
keras==3.9.2
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.0.9
numpy==2.1.3
oauthlib==3.2.2
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pillow==11.2.1
protobuf==4.25.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
typing_extensions==4.5.0
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
whitenoise==6.9.0
wrapt==1.17.2
